+++ core/api_container/server/service_network/user_service_launcher/user_service_launcher.go
@@ -12,6 +12,7 @@ import (
 	"github.com/kurtosis-tech/kurtosis/api_container/server/service_network/service_network_types"
 	"github.com/kurtosis-tech/kurtosis/api_container/server/service_network/user_service_launcher/files_artifact_expander"
 	"github.com/kurtosis-tech/kurtosis/commons"
+	"github.com/kurtosis-tech/kurtosis/commons/object_labels_providers"
 	"github.com/kurtosis-tech/kurtosis/commons/object_name_providers"
 	"github.com/palantir/stacktrace"
 	"net"
@@ -25,6 +26,8 @@ type UserServiceLauncher struct {
 
 	enclaveObjNameProvider *object_name_providers.EnclaveObjectNameProvider
 
+	enclaveObjLabelsProvider *object_labels_providers.EnclaveObjectLabelsProvider
+
 	freeIpAddrTracker *commons.FreeIpAddrTracker
 	
 	shouldPublishPorts bool
@@ -35,8 +38,8 @@ type UserServiceLauncher struct {
 	enclaveDataVolName string
 }
 
-func NewUserServiceLauncher(dockerManager *docker_manager.DockerManager, enclaveObjNameProvider *object_name_providers.EnclaveObjectNameProvider, freeIpAddrTracker *commons.FreeIpAddrTracker, shouldPublishPorts bool, filesArtifactExpander *files_artifact_expander.FilesArtifactExpander, enclaveDataVolName string) *UserServiceLauncher {
-	return &UserServiceLauncher{dockerManager: dockerManager, enclaveObjNameProvider: enclaveObjNameProvider, freeIpAddrTracker: freeIpAddrTracker, shouldPublishPorts: shouldPublishPorts, filesArtifactExpander: filesArtifactExpander, enclaveDataVolName: enclaveDataVolName}
+func NewUserServiceLauncher(dockerManager *docker_manager.DockerManager, enclaveObjNameProvider *object_name_providers.EnclaveObjectNameProvider, enclaveObjLabelsProvider *object_labels_providers.EnclaveObjectLabelsProvider, freeIpAddrTracker *commons.FreeIpAddrTracker, shouldPublishPorts bool, filesArtifactExpander *files_artifact_expander.FilesArtifactExpander, enclaveDataVolName string) *UserServiceLauncher {
+	return &UserServiceLauncher{dockerManager: dockerManager, enclaveObjNameProvider: enclaveObjNameProvider, enclaveObjLabelsProvider: enclaveObjLabelsProvider, freeIpAddrTracker: freeIpAddrTracker, shouldPublishPorts: shouldPublishPorts, filesArtifactExpander: filesArtifactExpander, enclaveDataVolName: enclaveDataVolName}
 }
 
 /**
@@ -94,25 +97,32 @@ func (launcher UserServiceLauncher) Launch(
 		volumeMounts[artifactVolName] = mountpoint
 	}
 
-	containerId, hostPortBindings, err := launcher.dockerManager.CreateAndStartContainer(
-		ctx,
+	containerName := launcher.enclaveObjNameProvider.ForUserServiceContainer(serviceGUID)
+	containerLabels := launcher.enclaveObjLabelsProvider.ForUserServiceContainer(serviceGUID)
+	createAndStartArgs := docker_manager.NewCreateAndStartContainerArgsBuilder(
 		imageName,
-		launcher.enclaveObjNameProvider.ForUserServiceContainer(serviceGUID),
-		dockerContainerAlias,
-		nil,	// User services won't run in interactive mode
+		containerName,
 		dockerNetworkId,
+	).WithAlias(
+		dockerContainerAlias,
+	).WithStaticIP(
 		ipAddr,
-		map[docker_manager.ContainerCapability]bool{},
-		docker_manager.DefaultNetworkMode,
+	).WithUsedPorts(
 		usedPorts,
+	).ShouldPublishAllPorts(
 		launcher.shouldPublishPorts,
+	).WithEntrypointArgs(
 		entrypointArgs,
+	).WithCmdArgs(
 		cmdArgs,
+	).WithEnvironmentVariables(
 		dockerEnvVars,
-		map[string]string{}, // no bind mounts for services created via the Kurtosis API
+	).WithVolumeMounts(
 		volumeMounts,
-		false,		// User services definitely shouldn't be able to access the Docker host machine
-	)
+	).WithLabels(
+		containerLabels,
+    ).Build()
+	containerId, hostPortBindings, err := launcher.dockerManager.CreateAndStartContainer(ctx, createAndStartArgs)
 	if err != nil {
 		return "", nil, stacktrace.Propagate(err, "An error occurred starting the Docker container for service with image '%v'", imageName)
 	}