+++ core/commons/enclave_manager/api_container_launcher_lib/api_container_launcher/api_container_launcher.go
@@ -1,0 +1,25 @@
+package api_container_launcher
+
+import (
+	"context"
+	"net"
+)
+
+type APIContainerLauncher interface {
+	// All launcher versions must implement this interface. This allows us to force developers, as much as possible,
+	// to handle backwards-incompatibility issues.
+	Launch(
+		ctx context.Context,
+		containerName string,
+		containerLabels map[string]string,
+		enclaveId string,
+		networkId string,
+		subnetMask string,
+		gatewayIpAddr net.IP,
+		apiContainerIpAddr net.IP,
+		otherTakenIpAddrsInEnclave []net.IP,
+		isPartitioningEnabled bool,
+		externalMountedContainerIds map[string]bool,
+		shouldPublishAllPorts bool,
+	) (string, error)
+}